requirements.txt
beautifulsoup4==4.3.2
ystockquote
blist==1.3.6
schematics==0.9-4
schema==0.3.0
Cython==0.20.1
numpy==1.8.1
python-dateutil==2.2
pytz==2014.2
Quandl==1.9.7
requests==2.2.1
six==1.6.1
PyYAML==3.11
dna==0.0.2
-e git+https://github.com/pydata/pandas.git@master#egg=pandas-0.13.0-27-gf7aeaeb
-e git+https://github.com/quantopian/zipline.git@master#egg=zipline-0.5.11.dev
